+++ filestorage/client_options.go
@@ -46,12 +46,12 @@ type ClientOptions struct {
 	// UID value to remap to when squashing a client UID (see
 	// identitySquash for more details.) If unspecified, defaults
 	// to `65534`.
-	AnonymousUid *int `mandatory:"false" json:"anonymousUid"`
+	AnonymousUid *int64 `mandatory:"false" json:"anonymousUid"`
 
 	// GID value to remap to when squashing a client GID (see
 	// identitySquash for more details.) If unspecified defaults
 	// to `65534`.
-	AnonymousGid *int `mandatory:"false" json:"anonymousGid"`
+	AnonymousGid *int64 `mandatory:"false" json:"anonymousGid"`
 }
 
 func (m ClientOptions) String() string {
