+++ objectstorage/object_lifecycle_rule.go
@@ -28,7 +28,7 @@ type ObjectLifecycleRule struct {
 
 	// Specifies the age of objects to apply the rule to. The timeAmount is interpreted in units defined by the
 	// timeUnit parameter, and evaluates against each object's Last-Modified time.
-	TimeAmount *int `mandatory:"true" json:"timeAmount"`
+	TimeAmount *int64 `mandatory:"true" json:"timeAmount"`
 
 	// The unit that should be used to interpret timeAmount.  Days are defined as starting and ending at midnight UTC.
 	// Years are defined as 365.2425 days long and likewise round up to the next midnight UTC.
