+++ objectstorage/transfer/multipart_manifest.go
@@ -17,7 +17,7 @@ type multipartManifest struct {
 }
 
 type uploadPart struct {
-	size     int
+	size     int64
 	offset   int64
 	partBody []byte
 	partNum  int
@@ -30,11 +30,11 @@ type uploadPart struct {
 // splitFileToParts starts a goroutine to read a file and break down to parts and send the parts to
 // uploadPart channel. It sends the error to error chanel. If done is closed, splitFileToParts
 // abandones its works.
-func (manifest *multipartManifest) splitFileToParts(done <-chan struct{}, partSize int, file *os.File, fileSize int64) <-chan uploadPart {
+func (manifest *multipartManifest) splitFileToParts(done <-chan struct{}, partSize int64, file *os.File, fileSize int64) <-chan uploadPart {
 	parts := make(chan uploadPart)
 
 	// Number of parts of the file
-	numberOfParts := int(fileSize / int64(partSize))
+	numberOfParts := int(fileSize / partSize)
 
 	go func() {
 		// close the channel after splitFile returns
@@ -47,7 +47,7 @@ func (manifest *multipartManifest) splitFileToParts(done <-chan struct{}, partSi
 		// Second go routine should start at 100, for example, given our
 		// buffer size of 100.
 		for i := 0; i < numberOfParts; i++ {
-			offset := int64(partSize * i) // offset of the file, start with 0
+			offset := partSize * int64(i) // offset of the file, start with 0
 
 			buffer := make([]byte, partSize)
 			_, err := file.ReadAt(buffer, offset)
@@ -69,8 +69,8 @@ func (manifest *multipartManifest) splitFileToParts(done <-chan struct{}, partSi
 
 		// check for any left over bytes. Add the residual number of bytes as the
 		// the last chunk size.
-		if remainder := int(fileSize % int64(partSize)); remainder != 0 {
-			part := uploadPart{offset: int64(numberOfParts * partSize), partNum: numberOfParts + 1}
+		if remainder := fileSize % int64(partSize); remainder != 0 {
+			part := uploadPart{offset: (int64(numberOfParts) * partSize), partNum: numberOfParts + 1}
 
 			part.partBody = make([]byte, remainder)
 			_, err := file.ReadAt(part.partBody, part.offset)
@@ -92,7 +92,7 @@ func (manifest *multipartManifest) splitFileToParts(done <-chan struct{}, partSi
 // splitStreamToParts starts a goroutine to read a stream and break down to parts and send the parts to
 // uploadPart channel. It sends the error to error chanel. If done is closed, splitStreamToParts
 // abandones its works.
-func (manifest *multipartManifest) splitStreamToParts(done <-chan struct{}, partSize int, reader io.Reader) <-chan uploadPart {
+func (manifest *multipartManifest) splitStreamToParts(done <-chan struct{}, partSize int64, reader io.Reader) <-chan uploadPart {
 	parts := make(chan uploadPart)
 
 	go func() {
